requirements.txt
boto3==1.26.61
botocore==1.29.61
openai==0.27.2
pydub==0.25.1
python-dotenv==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
streamlit==1.17.0